// DataAnalyse/types.ts
export interface DataRow {
  id: string;
  [key: string]: any; // Permet des champs dynamiques
}
// ...
export interface PivotJoin {
  id: string;
  datasetId: string;
  joinKeyPrimary: string;
  joinKeySecondary: string;
}